more updates
git-svn-id: file:///usr/local/svn/cpplib/trunk/cftal@1025 18305beb-b1e2-0310-bcd8-b14b5419b1d7

+++ d_real.h
@@ -116,7 +116,7 @@ namespace cftal {
             const float split_scale_down = 1.0f/16384.0f;
             // 2^14
             const float split_scale_up = 16384.0f;
-            cmp_result_type is_big(fabsf(a) > split_threshold);
+            cmp_result_type is_big(std::fabs(a) > split_threshold);
             float temp;
             if (unlikely(any(is_big))) {
                 a*=split_scale_down;
